Allowing users to enter a negative number at the begining of the program. Solves the issue of hitting multiple operation buttons in a row.
1:
Adds a boolean called 'Cleared' that is true when the program starts/clear function ran.

It becomes false if any number button is pressed.

If cleared is true, we insert a negative symbol into the text box when the minus button is pressed.

2:
Fixes weird issues when pressing multiple operation(+-*/) buttons in a row.

Now when doing so, it only worries about the last one pressed, instead of the first one.

3:
adds stopgap solution to formatting issue. needs more work.